Move test message types to `sandpack-client`
